Move optimizer code to separate subpackage
