+++ baybe/optimizers/__init__.py
@@ -1,0 +1,9 @@
+"""Optimizers."""
+
+from baybe.optimizers.base import OptimizerProtocol
+from baybe.optimizers.basic import GradientOptimizer
+
+__all__ = [
+    "GradientOptimizer",
+    "OptimizerProtocol",
+]
