+++ baybe/optimizers/base.py
