--- baybe/recommenders/pure/bayesian/botorch/optimizers/__init__.py
@@ -1,7 +1,0 @@
-"""Optimizers."""
-
-from baybe.recommenders.pure.bayesian.botorch.optimizers.basic import GradientOptimizer
-
-__all__ = [
-    "GradientOptimizer",
-]
